refactor: isolate Context controls and layer restoration
