+++ scripts/package-boundaries.json
@@ -369,5 +369,26 @@
       "src/ui/cctvPresentation.js"
     ],
     "external": []
+  },
+  "context-controls": {
+    "exports": [
+      "./ui/context",
+      "./ui/context/policy"
+    ],
+    "modules": [
+      "src/ui/context.js",
+      "src/ui/contextActions.js",
+      "src/ui/contextBindings.js",
+      "src/ui/contextControls.js",
+      "src/ui/contextLayerChanges.js",
+      "src/ui/contextPresentation.js",
+      "src/ui/contextSession.js",
+      "src/ui/contextSubscriptions.js",
+      "src/ui/contextTransactions.js",
+      "src/contextModePolicy.js",
+      "src/rightRailPolicy.js",
+      "src/ui/panelRailGeometry.js"
+    ],
+    "external": []
   }
 }
